DQN learning framework basically done.
We've add code to basically carry out the algorithm proposed by the paper.
The net, however, has not been built yet, so there's no way to test if the framework works fine.
Also there are two remaining feature not implemented : RMSProp, and the use of two separate networks for learning and target.